gnosis-keyper: move config to package
