+++ rolling-shutter/keyperimpl/gnosis/config/config.go
@@ -1,4 +1,4 @@
-package gnosis
+package config
 
 import (
 	"io"
